refactor: 자동차경주 클래스 패키징
