+++ src/main/kotlin/racingcar/domain/Location.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 data class Location(
     val value: Int
