+++ src/main/kotlin/racingcar/domain/MovingStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 fun interface MovingStrategy {
     fun isMovable(): Boolean
