+++ src/main/kotlin/racingcar/domain/RandomStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import kotlin.random.Random
