+++ src/main/kotlin/racingcar/dto/RacingGameRequest.kt
@@ -1,3 +1,3 @@
-package racingcar
+package racingcar.dto
 
 data class RacingGameRequest(val inputCarCount: Int, val inputLaps: Int)
