+++ src/main/kotlin/racingcar/dto/RacingGameResponse.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.dto
 
 data class RacingGameResponse(
     val lapReports: List<List<Int>>
