+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.view
+
+import racingcar.dto.RacingGameRequest
 
 class InputView {
     companion object {
