+++ src/test/kotlin/racingcar/CarTest.kt
@@ -2,6 +2,8 @@ package racingcar
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
+import racingcar.domain.Car
+import racingcar.domain.Location
 
 class CarTest : DescribeSpec({
