+++ src/test/kotlin/racingcar/LocationTest.kt
@@ -6,6 +6,7 @@ import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
+import racingcar.domain.Location
 
 class LocationTest : DescribeSpec({
